ios: register the script message handler as 'lok', not 'lool'
Signed-off-by: andreasma <maand@gmx.de>

With the commit 4063dcdc7305177b167bd20b009b1dfede8b25ec in the
LibreOffice core the naming for the message handler has changed. It
changed to 'lok'. Thus this change has to be followed here to fix
a break, introduced with a former change of the naming from 'lool' to
'cool'.

Change-Id: Ie298a6253522a5a6a9a878a249062eacf576e06f