wsd: disable mounting when unmount fails
Signed-off-by: andreasma <maand@gmx.de>

Adapt commit 5c7d94206f4bd57a4c78affa096c1f12acc89d5c from Ashod
Nakashian to LibreOffice Online.

It seems that on some systems it is possible
to succeed in mounting, but fail to unmount.

This is proving very problematic and it is
best to detect unmount failures and disable
mount-binding altogether.

Change-Id: Ie00746ad70e4a92458479cb6750fbb265ac66308