wsd: test: refactor client connection management
Adapt commit 5892bcba1ea04de69aa3f530cdb1d4239a096507 from Ashod
Nakashian to LibreOffice Online.

The client connection management had nothing to
do with the WopiTestServer and it was there
for convenience. The issue was that it was
not available to tests that do not need
the custom WopiTestServer and depending
on it would make them unnecessarily
complicated.

This adds a new intermediary class to manage
client connections in tests. The reason
this logic didn't move to UnitWSD or even
UnitBase is simply because the client
connection logic depends on the helpers
namespace and other test-specific code
that isn't available in LOOLWSD. In short,
UnitBase and UnitWSD are primarily an
interface that LOOLWSD and co depend on
and cannot contain test-specific logic
that isn't linked to coolwsd.

We also move the UnitWebSocket helper class.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I678bbe6166d542480f77b8c709361482ee8f3114