wsd: remove default parameters in Session
Adapt commit a56721f9cfbc7db7cc44db69d44d8c3d845397b9 from Ashod
Nakashian to LibreOffice Online.

And make arguments explicit for a setReadOnly
and setAllowChangeComment.

Default parameters are useful for backward
compatibility, when changing an API. But
they provide no benefit when the function
in question is new and is called a couple
times anyway. Code is harder to read with
default parameters because one has to
either guess (bad) or check (wasteful).

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I86a27077b05516ed8a5e3610e975de3889dc78c9