wsd: test: merge the results of all tests
Adapt commit 956f37c7a750d2fdd370f11350493b40b0ea2201 from Ashod
Nakashian to LibreOffice Online.

Since we can have multiple tests in a single
unit-test .so, we need to merge the results
of each test into a single result that is
used when we exit the process and flag for
failure or success.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I9609a8eaac2440f81893309a81de04cac0135744