We use String.locale unconditionally so make sure it always gets set
Even if the URL did not have a query parameter called "lang".

And yes, we should stop using String.locale for that, and instead add
our locale property to window.app, which seems to be the place for our
code to store various global state. Using String.locale is hugely
confusing. I naturally thought it was some built-in JavaScript thing
and was very confused when it was undefined and I couldn't find any
documentation on it.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: I549d3dac1be0991f87f1ca3aa3f147b081e898b1