wsd: move isWritable to Session and rename
Adapt commit 2b7e2bda963a0a14708ca1b57586e36fbcf99ea1 from Ashod
Nakashian to LibreOffice Online.

This helper is only true if editing of the document
is enabled and not whether or not the document is
writable in storage. Since it unifies isReadOnly
and isAllowChangeComments, which are in Session,
there is no advantage in having it in ClientSession.

It is now renamed to isEditable in Session.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I74dc942e8fb695eddaec78951dc313ae4a0750e2