calc cell cursor: get the layout and sizing right.
Adapt commit 616197d7a1de33e987179f83ab369fb4867c428d from Michael Meeks
to LibreOffice Online.

The cursor should minimally obscure the cell content - so the border
should go outside and over the cell border, and the white contrast
line inside it.

Similarly the autofill handle should not obscure the number in the
cell you're trying to read - so move it out and down to a more
familiar position, shrink the size somewhat, and correct the
white boarders around it too.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I0048159212ef8444c7d0a99befa4a6bddf801642