wsd: more accurate possible-modification flagging
This improves the flagging of possible-modification
in cases where there might be a race between the
arrival of the ModifiedStatus flag from Core
and the closing of the document. This can happen,
for example, when the user makes an edit and
immediately closes the document. In these cases
we want to be able to assume modification based
on the last activity from the user that may have
modified the document.

The issue this patch fixes is that we were too
aggressive in our assumption of modification,
because it was better to be wrong and try to
save rather than not. However, this also set
the IsModifiedByUser header to the PutFile
request when uploading the document to storage,
which was not accurate or acceptable.

Adapt commit 318f3f6eb82178872d3fb0aad710a7430a004490 from Ashod
Nakashian to LibreOffice Online.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: Ia3e1e98eea2aded4a380942212d8da17a2a6a5a5