Update javascript package d3 to version 7.8.2
The update to current version of the package d3 solved
some high severity vulnerabilities of the latest version
of the 6.x.x version. Because the new version changed
to module structure it needs an adaption in the
javascript code of the Admin part of LibreOffice
Online (done in main-admin.js).

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I8deaade00a18e126e78184987eee5cd674307afc