Modify the Info.plist to ensure that CFBundleVersion is always increm…
…ented

AppStoreConnect requires each upload, whether it is released or not, to be
higher than the previous successful upload's CFBundleVersion. So this
script sets the CFBundleVersion to the first and second components of
CFBundleShortVersionString and the UTC timestamp of when this script was
run is appended as the third component (or the second component if there
isn't a second component in CFBundleShortVersionString).

Adapt commit 3bd8dd063f94158ec32f90620bdeee4a24187126 from Patrick Luby
to LibreOffice Online.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: Ifc73b0b59f5a1ede3baac5c5a0ed8785890a7b5c