jsdialog: handle multiple tab controls inside dialog
It was possible to use only one tab control.
Now useSetTabs parameter for builder is introduced
which defines if we want to use custom placement
of tabs handled by the parent container or put
tabs in the place where they are defined in JSON.

Custom placement is needed in mobile-wizard and notebookbar
as for them tabs needs to be placed at the top bar.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I6420e303de248058a7a94ead47b6672b717360df