JSDialog overlay: Use CSS rules for positioning.
Set position rule and some others of the overlay to absolute. So we can use it both for popups and documentareapopups.
Set max-height of the pop-up so it doesn't look ugly when the height of the popup is bigger than the window height.
Modify JSDialog for the new behaviour.
I also split the code into functions for easier maintaining while updating.

Add special rule for comments. They are not contained in the container.

Adapt commit a2d666d53adcb23e60ed61328e8e3fe44fdc1af5 from Gökay Şatır
to LibreOffice Online.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I32eeca97ad4fc567ac7d13b7b711383d379de76b