wsd: do not warn of unassociated Kit after unloading
After unloading a document, the ChildProcess instance
may be destroyed by the time that we process the
internal prisoner connection. In that case, we
can't find the ChildProcess so we assume it was
a premature death and issue a warning.

By tracking whether or not a PrisonerRequestDispatcher
had an associated DocBroker, we can supress warnings.

Adapt commit 58566083d51e5a4bbd8c4f5101b4e93bea065c92 from Ashod
Nakashian to LibreOffice Online and add missing lines.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: Ib9ea7bc26280c65ef7d9704ba022bf6fd93c57c2