wsd: announce the debug URLs to the user after server runs
The port used for client connections does change when
we cannot connect on the default/first one. There is
no point in announcing the default when we don't
yet know the actual port, and similarly it's not
helpful to do so before the server is even started.

We also move the remote-font-config thread to an
earlier point and merge it with another #if !MOBILEAPP.

Adapt commit 9dd0636bcd424658e248e2e1ed83ff28b1160e7f from Ashod
Nakashian to LibreOffice Online.

Signed-off-by: andreasma <maand@gmx.de>
Change-Id: I7b399dba0e97d0d50148a7b9c7edfc0dcb6f0521