Related: #6423 restore Pre-fetching next or previous slide tiles
originally done with:

commit 067381f0b273a4c623715d022ce54a698bae9bb7
Date:   Wed Dec 2 16:46:40 2020 +0300

    Pre-fetching next or previous slide tiles

but didn't make its way into the rework of

commit a15597de9c88d6f52cfce6808b83a06a3c089960
Date:   Tue Mar 30 13:16:14 2021 +0300

    Merge GridLayer+TileLayer+CanvasTileLayer

but I think this is broken since even earlier.

At this point the next/prev slide tiles are fetched, but dismissed
and not cached properly yet

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: I3db2f5b228200e6cfb02c005bb8d2b2d659459c7

Fix #6423 not requesting a large enough range in pre-cache

request in pre-cache the same range that we end up requiring later.

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: I5d6632ce6dbf238578daa988136983097188ce1d