A11yTextInput: empty selection at the beginning of a list entry
When there is an empty selection at the beginning of a list entry and
a <backspace> is typed, the <backspace> doesn't delete only the empty
selection as it occurs in a normal paragraph.
So we need to avoid to send 2 backspaces.
Since both a <backspace> and a <delete> removes the empty selection we
always send a <delete> followed by a <backspace> when a <backspace> is
typed and we need to remove an empty selection.
This solution works both for normal paragraph and for list entries.

Signed-off-by: Marco Cecchetti <marco.cecchetti@collabora.com>
Change-Id: Idae144b65ed04ac28c602ee2155d337369c534fe