wsd: trim memory on idling
When all users idle their activities
for at least 10 seconds, we take the
opportunity to trim the memory footprint.

We are careful not to issue a subsequent
trimming too soon. A minimum of 30 seconds
needs to pass to trim again. This should
give a decent balance between performance
and reduced memory footprint.

Change-Id: I1c8b2c4a265c9aa0ac5033762695706ad970f11e
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>